+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [0.1.0] - 2025-09-23
 
-- Initial release
+- Index a model by including `Bivy::Indexable` module.
+- Access indexable models through `Bivy::Indexable.models`.